GUI walletimport RPC call added
most of the code was in bitcoingui. i tried to make it clean and used
QInputDialog instead of making a model for the wallet just to create the
dialog. moreso, that code would be out of rpcdump space but still using
code mostly from rpc. i decided neither is really the best option and
used only what i needed from rpcdump.