Updates that were missed during the merge from blackcoin master
Remove -logtodebugger

`-logtodebugger` is a strange, obscure, WIN32-only (mostly MSVC) thing.
Let's clean up the options a bit get rid of it.

test_bitcoin was using fLogToDebugger as a way to prevent logging to
debug.log. For this, add a boolean (not exposed as option) fLogToDebugLog that
defaults to true and is disabled in the tests.

Conflicts:
	src/init.cpp
	src/test/test_bitcoin.cpp

remove zerocoin leftovers

fixed to main.cpp

fixed to main.h

fixed to main.h

fixes for code updates